Add support for configuring virtio random number generators
libvirt has the capability of passing through random numbers from either
a random number generator on the host or an EGD-compatible socket. Add
code to configure that in nova.

Latest libvirt requires a different XML format for the virtio RNG.
Add support, including some hard-coded details for the EGD device -
we'll fix that up once we decide which options should be offered for
configuration.

Change-Id: I15cbb379a3f4ff3cd7b59b226b970c34781601c1