REMOVE THIS IF DB IS SYNCED FROM OSLO-INCUBATOR
Catch deadlock exceptions in .all() queries

This patch is to allow deadlock retry's on all queries. It has not
been merged upstream. If it is merged upstream and backported this
should be removed as it will likely conflict.

Original patch: https://review.openstack.org/#/c/75802/3/

We haven't seen the specific deadlock we were seeing since merging
the parent of this commit, but we are pulling this in just in case the
deadlock is still possible under some scenarios.

Change-Id: Id9d9bf72dff0ca8602a192fc74ab1e87b57d5c5e