Fix ownership checking in get_networks_by_uuid
The code was elevating context before requesting networks which
means that the project_only code is skipped and all networks
could be retrieved. This means that the default networks returned
by FlatDhcpManager could include networks that belong to other
projects.

This fixes the issue by requesting the network without elevating
the context, re-enabling the proper project checking. It includes
tests to verify that the proper exception is raised when an
illegal network is requested and that the context has not been
elevated by the compute manager.

Partially-implements blueprint better-support-for-multiple-networks

Change-Id: Icd3bc521003725cc3da9875dfd6532d5c5524f43
Closes-Bug: 1331092
(cherry picked from commit 1e42100a10dc3966581264bd55e4703c6e45ff81)